defis/19_Professeur_Dumbledore/requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
soundfile>=0.12.1
torch>=2.2.0
numpy>=1.24.0
librosa>=0.10.1
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.1
gTTS>=2.5.1
